[TASK] Remove FED dependency
FED is used only for its ViewHelpers related to auth. This will probably
be removed entirely at some point by utilizing the selfmade setup that
originated from Appointments, but in the mean time we can simply replace
FED with VHS and be rid of it, as it contains interchangable versions of
the related ViewHelpers.

Change-Id: If9fc5abde3e0f3fe14cb46d09982d6b63a9afff3