[TASK] Make Category-File/-Link relations bidirectional in TCA
The category could only be set in TCA from File or Link records. This
limits us in several aspects, like not having the option to change files
and links assignment from the Category record, or in FE having to rely
on the respective repositories and findByCategory() methods rather than
having (lazy) objectStorages at your disposal right away. This commit
changes that.

Change-Id: I58068b3056e88891c9c99a74dc8d22fd526a127a