[BUGFIX] FE adds don't alter category counters
The file and link properties of the category domain don't get updated
with the correct count, if a file or link is added through FE. This is
because originally, the category-relation was only present from files
and links, thus they're the only instances being updated from the
file-/link-controller. Since the relation is bidirectional, we can make
the counters remain consistent by also adding the files and links to the
category instance from the file-/link-controller, updating both sides of
the relation.

This change helps keep the counters from the category-side reliable,
which is necessary once we start displaying the category-counters
somewhere without initializing its still-lazy file/link properties. When
(if) we change the user interface to a single page application through
ajax calls, we get to benefit even more from this being consistent.

Change-Id: I249074c802f50754b7eea0d9d48f8a6ca0c9e099