[BUGFIX] Superuser doesn't work anymore
Since the change from FED to VHS and recordType-agnostic feuser records,
a superuser is no longer recognized as such. It seems VHS does an
splhash check with an original extbase model, which will always be
different than the fileman model.

Seems it's already time to apply the autogroup/appointments user
services in fileman then, and get rid of VHS as well. This would fix the
bug and any issue that might be related to it.

Change-Id: I5533fca3ad7c23653103446fd5845c905bf69544