[TASK] Clean up field id's and css
Fields have very general ID's, like "submit" or "title". There is a
relatively high chance these could occur in other parts of the webpage.
By prefixing them with the extension key, we're ahead of that. Also, CSS
is based for some part on ID's, but we could easily utilize classes of
course.

Change-Id: Ia2ba295b1095e582341648b14f09481f9798812f