[TASK] Make files and links order alphanumerical
As with categories, a wish for a static sorting was desired, hence an
alphanumerical sorting is now applied on repository level.

However, because of the "alternative name" mechanic of both files and
links, there were two fields which both had the potential of being the
display title. Due to this, I've changed the mechanic so that the
"alternative name" is always used, and if empty on create/update, it is
automatically filled. There are a couple of sidenotes to this:

1) In TCA, the easiest route was to simply make the field required. We
might want to change that later.
2) It would be nice if in addition to the current mechanics, the input
field is automatically filled through javascript, as this would make
things more clearer to the user.

Both should be dealt with in a later commit, perhaps?

Change-Id: I5dae019bcfa29d5593d1b04639ec1e482ab8fc76