[TASK] Replace jQuery UI script
The only reason we have the script is to display a progress bar, but for
our purposes, we can get away with an extreme minimum of very easy code
that the jQuery UI is practically bloat. Replacing it with a few lines
of our own js and css renders the jQuery UI js and css useless.

Change-Id: Iad8ebf911b4132ad6b5858e812b2459bb707a70f