[BUGFIX] Add/Remove upload fields picks wrong indices
Adding a new field seems to always pick the first to copy, which results
after the second field in duplicate indexes because the JS counter can
nog longer find its previous value to replace in the copy.

Removing a field happens backwards: not the last, but the first field is
removed.

Both issues have a common cause: the jQuery selector used with its
prevAll() function is set to :last, but in prevAll() context, this is
the last counting backwards.

Change-Id: I7daf62d81f5c97bc06ba2c03461500a5310857d5