[BUGFIX] File upload form 'formObjectName' errors @ TYPO3 6.x
All fields are replaced by the same error message when
rewrittenPropertyMapper is turned off. Turns out that the
ViewHelperContainer can't get the form class by its original
non-namespaced name, and fileman is affected by this because it uses
custom form ViewHelpers which retain the original class-names. A
version_compare is all we need.

Change-Id: I8e4d8f55dae8000fc7ef554f93826ccc7ff12855