[TASK] Replace fileUri field with uploadData field
Fileman validation checks whether file->fileUri is NULL or not, to
determine on validation issues if this property was clean already, but
the rewrittenPropertyMapper will make it an array regardless, breaking
these checks and requiring several additional checks. To force
consistency between propertyMappers, fileUri is to be replaced as
upload-field.

Change-Id: I8927868c70d201d0436490c2384208dbc5ce1658