[BUGFIX] Denied file-type error doesn't clear upload-field
The upload-field is replaced with the last successfully uploaded
file-name, even if the file doesn't meet validation requirements. The
upload field needs to be retained in those cases. An easy fix would be
to clear the file-name in the upload-data through fileService, once a
uploadData/fileUri validation error has ensued.

Change-Id: Ibec5f1eed029f06de8b1531da32b7865dfd01c1e