[BUGFIX] TYPO3 Exception after validation error with successful file
When a file is successfully uploaded, but a validation error ensues for
different reasons, the findSubstitutes() mechanism kicks in to ensure
the file can be moved successfully after validation succeeded. It turns
out that in linux servers, the temp-files have no (or not always) a .tmp
extension, yet they are validated with a pattern that requires them to.
This results in not setting a substitute, and no index available,
destroying the entire form when processed again. The Form VH is the
first to throw an exception, because of a missing index resulting in
formFieldNames that have an empty array somewhere inbetween, e.g.
files[file][][fileUri].

The fix, is to loosen up the validation for the temp file, to not
necessarily include the extension.

Change-Id: I0000000000000000000000000000000000000000