[TASK] Explicitly disable rewrittenPropertyMapper
The rewrittenPropertyMapper changes many things in how extbase handles
properties and errors for properties in the fluid form VHs. It's quite
some work to accomodate this. It's enabled by default in 6.x for
extensions that don't disable it explicitly.

Since disabling the rewrittenPropertyManager is supported until 6.3, and
we currently aim for 6.2 compatibility, we're disabling it for now.

Change-Id: I4c480bdd456dd064a7b3f70695ca7efee9e58ddf