[TASK] Make Textfield VHs 6.2 compatible
Between TYPO3 6.1 and 6.2, the placeholder attribute for the Textfield
VH was moved from a render-argument to one initialized in the
init-method. To be compatible with both versions, we need te rename the
placeholder argument and simply pass it along the parent for <6.2, while
registering its value as the placeholder tag-attribute for 6.2.

Change-Id: I1f20cde52dbcb09b504e82c1bd6df05d01f1f5d8