[TASK] Make hacked ValidatorResolver 6.2-compatible
The 6.2 ValidatorResolver adds "Collection" validators by default to
Storages, but this messes up our own validation-hacks that we require
for reliable pre-6.2 validation. We can solve this by removing the
entire addition from the build-conjunction process.

Also, the 6.2 ValidatorResolver conflicts in argument-count for one of
its internal methods in versions prior, which we utilize for our
dontStoreConjunction mechanism (one of the above mentioned hacks). There
needs to be an exception to add support for the 6.2 way.

Change-Id: I60a63de963012fd0c9aad347a8c0ab67647fbe0e