[BUGFIX] Files can be made invisible by clearing the category field
When editing a file, clearing the new category field and saving, will
leave the file in limbo. Afterwards, the file can only be fixed from the
backend. Preferably, there should be a NotEmpty validation applied to
the file.category property so a file can never be saved with less than 1
category. It only needs to be tested in TYPO3 6.2, as we're moving away
from 4.5 support anyway.

Change-Id: Ibe308b20916f9197de8f1f8fb2e973a0a10f82cb