[TASK] Add CSRF-protection
Copy the CSRF-protection implementation from my appointments ext. I'd
like to wait for it to mature a bit in implementation, but a client
needs it NOW. A pity, considering how much work it still is to get it
running.

Immediately also replace DELETE links with DELETE forms.

There's just one issue I found: when a token fails for a delete-category
on a subcategory, the subsequent category-argument overrules the
original, placing you inside the category you tried to delete. This
issue should probably be dealt with when everything matures enough.

Change-Id: I77ed991f620cf163e32c427cf7ba3e115133df7d