[TASK] Add JS upload-method
We're stuck with php-specific upload methods, and all limitations that
come with it. This means we can never upload large files as we'll get
timeouts, regardless of php.ini settings.

One way to get past these limitations, is by passing the file in chunks
via separate XHR requests, possibly POST or PUT, of which the latter is
known to bypass several restrictions more than post.

HTML5 offers support for several file-specific API's, e.g. the slicing
of files, which we can use in tandem with XHR PUT requests.

Change-Id: I2fa2172ebad4e04d522082f1a79fe7f747ba40d7