[BUGFIX] Subcategories don't sort alphabetically
Currently, subcategories are shown by utilizing category.subCategory.

The subcategory relation sits in an MM table with its own sorting
property, but this sorting property is what determines the order of
category objects in category. It turns out I can overrule this via TCA,
but the desired sorting order is only achieved when saved in the
backend, via the TCA record. Herein lies another problem: subcategories
will for the most part be created on the frontend, and then the desired
sorting order isn't in effect until someone goes into the backend again
to manually save the new order.

So we need to provide additional sorting. We could do this via array
sorting logic, either in the controller or in a sorting viewhelper. But
both pollute the code compared to simply retrieving the subcategories
via the repository, which is already set to retrieve categories in the
desired sorting order. So for now, the latter has my preference.

Since the category.subCategory property is lazy, the subcategories won't
be retrieved from there with a query, until that property is accessed.
So with our preferred solution, we still only produce one query in
retrieving them.

Change-Id: Ifaf8845316e60444dbab7b5bc9382d7fc0e15808