[BUGFIX] Category admins that are not part of group aren't shown
If a category has a parent that is no(t) (longer) part of the
possibleOwners group, his name isn't shown despite him being the actual
owner. A first user on the list will be selected then, which could lead
to confusion and/or unintended assigning of admins.

Also the current user should ALWAYS be selectable, even if he isn't the
owner and not part of the ownergroup. One use case for this is when
superUsers are not part of the normal possible owner group.

Change-Id: I5fe739007914877601d272d3ce929b3d4681ae6f