[BUGFIX] Category->feUser is still lazy despite heavy use
Since the introduction of the category-owner/admin, the feUser property
is being used on every request now. We can shave off a a bit of a
performance-impact by making it un-lazy, as well as prevent the errors
currently issued due to our not taking into to account of the
LazyLoadingProxy object.

Change-Id: Ib7c1aa02a197af8838011ec0a44af2f47bb163b8