[BUGFIX] Files can be placed outside of any category
Doing so leads to unreachable files. On the long term, we might want to
do a File->list action as leading, even on page 0. And this might make
Category->list completely unnecessary.

But for now, the extension is not there yet. So we have to prevent files
from being placed outside of any category.

Change-Id: I68f35bb8448395a583b9fcf6a52e0c9e8cd0178e