[BUGFIX] Cannot remove file via search
Removing a file via search does not pass on a category, which was
required to first remove said file from category before detecting if it
belonged to any remaining categories.

It is therefor never removed from any category, and thus never unlinked.

An acceptable solution is to regard a deletion from search as the
suggestion that the file needs to be removed regardless of the number of
categories it is in. This means that if deletion happens with $category
= NULL, just remove it from ALL its categories and continue unlinking.

Change-Id: Ieee4a9ab6bfa1d98f91545ddbef0fcf57bacab1b