[BUGFIX] Progressbar just hangs on errors
If there is an error during upload, the visitor is not informed about it
and the progress bar sits there, just being pretty (no, not really).

There should at least be a visual confirmation and a tip surrounding the
issue, i.e. an error message that suggests to try it again after
reloading the page, or to contact the webmaster.

Change-Id: I3b75387a437a24c97e8ea0ab62eccbbbaf636faa