[BUGFIX] Deleting a file-entry in-between messes up XHR uploads
The issue is that file-entries get an index following the last one, but
removing a file-entry does not rectify the missing index. So when XHR
uploading replaces the fileupload field with a hidden field that has a
continious bunch of indexes, there is a mismatch. The fileUri will be
placed in what is supposed to be a non-existing index, while its
metadata exists in the correct index field, effectively creating the
expectation of 2 file-uploads instead of one, and failing validation as
well as setting up rectifying the form afterwards for issues by
combining wrong metadata.

We already have a getIndex() method in jslib, so just use that to detect
the uploadfield's actual index and using that for the hidden field, and
the problem should be solved.

Change-Id: I7f9cff1b27ec3b18a70964ca289860c99412d20c