[BUGFIX] JS upload filenumber does not always match backend
If you forget to pick one or more files AFTER a picked file, the jslib
seems to erroneously replace the upload field as if it contained a file.

This causes the issue that the property mapping error that ensues, can
not find the number of expected $_FILES parameters, causing a mismatch
and therefore losing anything you've filled in those file's extra
fields. Furthermore, it makes the process somewhat unstable if files
unexpectedly drop out of the equation.

Change-Id: Ic997e1a9c391240e9280e9bdf6b914d957d9da01