[BUGFIX] edit from search can break followup actions
A search action can land you in the File::list action without a category
parameter after editing one of the search results, which breaks any
subsequent action from its template that need a category. This is
because the search action is devoid of any category parameter. by itself
and therefore can't pass one to edit. An easy solution is to use the
first category of the file as a substitute.