[TASK] Replace inline fluid conditions with older syntax
Recent developments in Fluid make a more streamlined syntax for inline
conditions available, and it works in TYPO3 v8 and v9. But not in one
specific client case for whatever reason (no custom fluid version there
afaik). So instead, I've replaced the syntax with the older syntax and
it works everywhere.